% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{execute_experiment}
\alias{execute_experiment}
\title{Execute experiment}
\usage{
execute_experiment(experiment)
}
\arguments{
\item{experiment}{is a list with the name of a file to be executed and a
config fields with ENV vars that must be set before the execution.

The language env var must be set so we can decide which language to use
when executing the experiment.}
}
\description{
Execute experiment
}
