% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{parse_config}
\alias{parse_config}
\title{Takes a bench.yaml file and lists all the experiments that need to run.}
\usage{
parse_config(file = "bench.yaml")
}
\arguments{
\item{file}{path to configuration file.}
}
\description{
Takes a bench.yaml file and lists all the experiments that need to run.
}
